finishes Collection API - closes #37
